feat: partition routed workers
